fix: changes to methods for getting users
The method for getting users now supports a custom filter.
Returns an array of mathcing users with their passwords omitted.
Filter is not required, if it is not passed, the method will simply return all of the users.
Full info about a user is available by /user/me which only returns an authenticated user