fix: notice adjustments
Added closedAt attribute to notices (for future use in reports).
Made it so a notice can only be closed with /close method (in order to set close date) and so it cannot be reopened.
Fixed a vulnerability that anyone could edit any notice by id (now you can only edit your own notices)