fix: reports - excessive db calls
Introduced a config file that can be used to store some global values/settings, which may be used in different parts of code.
P.S. it is probably better to use environment variables, but since we don't dockerize the app, a file should cover for it.

Rewrote reports methods so they only call the db once and handle the results themselves.