fix: refactored GET request bodies
Some of the GET requests across the app used body as means of passing parameters and data to them. This way the content of request body affected GET methods response.
This behaviour does not follow common worldwide practices nor does it comply with semantics of a GET request: https://stackoverflow.com/a/983458
In addition to that, in some cases one migth not even be able to add a body to a GET request.

All of the GET methods were refactored so that the request body does not matter to the response. Some parameters must now be passed in query.
Affected methods:
- /report/periodic - 'to' and 'from' dates must now be passed in query
- /notice - a custom arbitrary filter may no longer be passed in the request body (one should use /filter/* methods instead)
- /notice/search - search query is now expected in 'q' query parameter (kinda analogous to the way it's done in google: https://www.google.com/search?q=sample+text)
- /user - arbitrary filter may no longer be passed in body; one may now use /user/search though (works the same way as /notice/search)